requirements.txt
aniso8601==8.0.0
astroid==2.3.3
click==7.1.1
colorama==0.4.3
dicttoxml==1.7.4
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.5.1
mccabe==0.6.1
pylint==2.4.4
python-dotenv==0.12.0
python-dotenv-run==0.1.4
python-simplexml==0.1.5
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.16
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.11.2
